--- client/requirements.txt
@@ -1,4 +1,0 @@
-streamlit
-streamlit-mic-recorder
-openai
-pandas
