+++ client/src/core_utils.py
